Rename nbs api and service to appointments
