+++ src/new-client/src/app/lib/services/api/appointmentsApi.ts
@@ -1,5 +1,5 @@
 import Client from '@services/api/client';
 
-export const nbsApi = new Client({
+export const appointmentsApi = new Client({
   baseUrl: `${process.env.NBS_API_BASE_URL}/api`,
 });
