+++ src/new-client/src/app/site/[site]/users/page.tsx
@@ -1,7 +1,7 @@
 import { Table } from '@components/table';
 import NhsPageTitle from '@components/nhs-page-title';
 import Link from 'next/link';
-import { fetchUsers, fetchRoles } from '@services/nbsService';
+import { fetchUsers, fetchRoles } from '@services/appointmentsService';
 
 type PageProps = {
   params: {
